Fixed Header and Nav Bar, Worked on Middle of Site
I fixed the disappearing issue with the header when in mobile mode, as well as editing the colors of the text in the header. I added colored background boxes for the history and variety sections, as well as a new font to act as the body text. I still need to align everything properly.